add naive upsample program
the theory behind upsampling:
to scale from 11025 to 44100, the integral scalefactor (N) is 4.
so each sample in source is converted into N output samples
as follows:
sample_t upsampled[N] = { source_sample, 0, 0, 0 };
1) use the source value
2) use N-1 0's.
3) apply some kind of filter on the 4 values

in the current implementation, 3) is skipped.
it seems to work well to upsample 22050 samples, but not ones with
lower samplerate.