multiple changes to make animations more accurate
previously, it could happen that the first frame of an animation
was only played for 1 instead of 4 ticks.

i introduced a new variable in gameobj to keep track on which tick
an animation started. additionally some code was refactored into
functions and shuffled around so we don't have off-by-ones.
for example, spawning an explosion in the main loop would cause
its step counter be incremented twice, or the first anim being skipped.

to distinguish a freshly spawned object from an older one, a special
anim initialization value was introduced so that anim_curr doesnt
get incremented twice in the same tick.