add unscramble tool
this was necessary to make the vehicles and turret sprites look
properly.
works as is for turret sprites.
for vehicles, they must be saved in a single vertical strip
16x2008 (from memory dumpoffset 259860), and one application
yields the 32 pix wide sprites (jeep), and another application
on the output yields the 64pix wide sprites (big green tank
and transporter).
for the medium sized gray/white tank, which is 48pix wide,
another modification is necessary.