refactor map handling
we now use a table to locate the current map screen.
this should make it cleaner to implement horizontal scrolling.
additionally we dont need to put empty mapscreens more than once
into the screens array, so it's a net gain in binary size.

the meaning off the map_off variable was also changed:
previously it was the y offset into the screen of the "current" map,
now it means how many y's of the topmost screen are out of the
visible range.
so map_yoff == 189 would mean that of top screen's 192 lines, 189
would be off, and only the last 3 lines displayed.
at 3, the 192 lines of the second screen would be displayed, which
leaves us 5 pixels off of the visible screens 200 lines.
so those 5 pixels would be taken from the third screen.

189:    183:    17:     0:
| 3     | 9     |       | 192
                |       |
| 192   | 191   | 175   |
|       |       |       |
|       |
|       |       | 25    | 8
                |       |
| 5