fix bug in right side scrolling
since the mapscreen_xoff var applies to the left screen,
we cannot continue increasing it when we hit the right-most one.
so in the moment we set the x screen to the right-most one, we must also
change the scrolldir.

this bug would have caused a segfault, wouldnt another bug in the
draw_map code (will be fixed in the next commit) have caused the left one
of the 2 visible map screens to be painted twice.