draw_map: fix horizontal drawing
the x_screen_iter var was always meant to serve as an offset into
the mapscreen table, however it was forgotten to add.
since the map screens in a horizontal scrolling area are always the
same, the bug was not noticed before.
now we could theoretically make "proper" maps for horizontal scrolling,
instead of simply using a plain one for all of the screens.