hackish OpenGL port
to use OpenGL properly, one has to know a shitload of things about it.
in order to keep it simple, each entire screen is converted from bitmap
format into an OpenGL texture. this is pretty lame and a real CPU hog.
we're now at 70% CPU at 2x scale, whereas the SDL video output uses 2-5%.
this was mainly done to test if OpenGL would allow us scrolling without
tearing. that doesn't seem to be case, at least not entirely.