add all missing enemy sprites (turrets, tanks, bunkers, ...)
apart from soldiers, this includes sprites of any hostile entity
in the game.

the only not imported sprites at this point are the borders
used by the weapon-purchase screen as well as the one which has
the different muzzleflashes used in the game and the blood
animation when an enemy soldier is hit.
the "boss" sprites are also missing, also the decoration for
the ingame player stats (lifes, jungle graphic, M.I.A text +
minimap pic).
the latter can all be easily taken from a screenshot, while the
former 2 are already ripped from the game and in preparation.