add spawnmap of australia (parts completed so far)
it's a hell lot of work to enter all the enemies.
thus only the first 3 screens are finished so far, despite spending
many hours on it.
the current configuration caused the UB fixed in the previous commit
when the first (and only) spawnentry for screen 4 got processed,
since it happens to contain 0 as the spawnline, which is equal to
the spawnline in the next chunk of memory if interpreted as spawn
entry.