add misc spritesheet
this one contains a lot of identified and unidentified small stuff.
- 2 different muzzleflash sprites (thick and less thick)
- blood splatters
- rocket, though missing the pointy head

and other stuff which seems to be unused in the game.
potentially some alternate bullet types that the developers played around
with back in the days.

the original spritesheet has 4 sprites per line, which are always equal
to the first one but shifted by some empty pixels to the right.

i'm adding this for completeness' sake, despite needing only the blood
graphics at the moment.
i plan to add the original 2 muzzleflash graphics, so we have the flash
alternating randomly.
but even then, only 1/4 of this sheet will be used.

in a heavily resource constrained environment (maybe NDS?) it may make
sense to omit the unused parts.
for the moment, we look quite good though, about 1MB total filesize
and RAM usage including all resources.

since there are all different kinds of sprites with different palettes
mixed up here, the palette does not match.

for muzzleflash and the original blood color, the following palette
can be used:

PRGB(  0,   0,   0),
PRGB(187,  85,  17),
PRGB(153, 153, 119),
PRGB(255, 221,  85),