add initial prototype for weapon shop
doesn't do anything except drawing stuff currently.
it uses the "old" weapon_sprites sheet which was taken from the
game's weapon sprites (but those that are displayed vertically
in the status area left and right), and were converted from vertical
to horizontal.
as a consequence, ammo is missing and there's only one grenade
instead of a bunch of them.
i couldn't find the horizontal sprites in the memory dump, so they
will be added from a screen dump later.