replace weaponshop sprite of M203 with MG42
1) we need a slot for the MG42
2) the M203 is identical to the M16A2 but has a grenade thrower
   attachment. however implementing that requires a lot of special
   case code which i'm unwilling to do. (all other weapons have
   exactly one ammo type, etc).
   also it is planned to make it possible to switch to regular grenades
   with a single keypress (unlike the original where you only can cycle
   through the available weapons). this makes the only advantage of the
   M203 the original had go away: there you could toggle the shot type
   of the active weapon with a single keypress.
   (well additionally the grenade shot allowed higher distance).

as a consequence, the grenade ammo for the M203 is now superfluous
and frees another slot that could be used for an additional custom
weapon (i'm thinking about adding a sniper rifle, preferably a
dragunov SVD). or it could be used to add authentic 7.92 mm mauser
ammo for the MG42.