add fail sound
the actual audio sample ripped from the game needed to be altered to sound
correct.

in order to do it, i put the sample into milkytracker, changed it to loop
"forward", played it as B3 (note) for 2 ticks and used EC0 effect on the
third tick to stop. then it was exported as wav via "Disk OP", and "save
as" in the song tab with 100% amp.