audio.c: make it possible to play pseudo empty tune
wave sounds are only played when music is playing as well, because
the sounds are mixed into the music stream.
having a separate logic and thread/lock state to deal with sounds
separately from music seemed like a lot of gratuitous effort,
so i just faked the functionality by playing a pseudo empty tune.

for some reason, when the empty tune starts playing, there's a short
unwanted noise.